secure-data-transmission/requirements.txt
numpy
opencv-python
Pillow
requests